int changed to double

+++ src/main/java/Operation.java
@@ -1,16 +1,16 @@
 public class Operation {
 
-    public static int sum (int a, int b){
+    public static double sum (double a, double b){
         return a + b;
     }
 
-    public static int sub(int a, int b){
+    public static double sub(double a, double b){
         return a - b;
     }
-    public static int mult(int a, int b){
+    public static double mult(double a, double b){
         return a * b;
     }
-    public static int div(int a, int b){
+    public static double div(double a, double b){
         return a / b;
     }
 }
